+++ i18n/giswater_en_US.ts
@@ -11012,6 +11012,14 @@
 			<translation>groupBox_4</translation>
 		</message>
 		<message>
+			<source>groupBox_5</source>
+			<translation>Extra filters</translation>
+		</message>
+		<message>
+			<source>tooltip_groupBox_5</source>
+			<translation>groupBox_5</translation>
+		</message>
+		<message>
 			<source>tab_connect</source>
 			<translation>Link</translation>
 		</message>